requirements.txt
beautifulsoup4==4.11.1
enum34==1.1.10
Pillow==10.0.0
soupsieve==2.3.2.post1
validators==0.21.2
